+++ app/javascript/client/home/HomeContainer.tsx
@@ -1,14 +1,9 @@
 import React, { FC, useContext, useState } from 'react';
 import { AuthContext } from 'client/contexts/AuthContext';
-import {
-  useLogoutMutation,
-  useGetUserForHomeContainerQuery,
-} from 'client/graphqlTypes';
+import { useGetUserForHomeContainerQuery } from 'client/graphqlTypes';
 import { AddPersonForm } from 'client/profiles/AddPersonForm';
 import { gql } from '@apollo/client';
 import { withRouter, Link } from 'react-router-dom';
-import { Wrapper } from 'client/common/Wrapper';
-import { NavBar } from 'client/nav/NavBar';
 
 gql`
   mutation Logout {
@@ -53,7 +48,6 @@ const InternalHomeContainer: FC<HomeContainerProps> = () => {
   if (!userId) window.location.href = '/login';
 
   const [newPersonFieldVisible, toggleNewPersonFieldVisible] = useState(false);
-  const [logoutMutation] = useLogoutMutation();
   const {
     data: userData,
     refetch: refetchUserData,
@@ -71,16 +65,8 @@ const InternalHomeContainer: FC<HomeContainerProps> = () => {
     </div>
   ));
 
-  const handleLogout = async () => {
-    await logoutMutation();
-    window.location.href = '/login';
-  };
-
-  const navMenuItems = [{ label: 'Log out', onClick: handleLogout }];
-
   return (
-    <Wrapper>
-      <NavBar dropdownItems={navMenuItems} />
+    <>
       <button onClick={() => toggleNewPersonFieldVisible(true)}>
         Add a new person profile
       </button>
@@ -88,7 +74,7 @@ const InternalHomeContainer: FC<HomeContainerProps> = () => {
         <AddPersonForm refetchUserData={refetchUserData} />
       )}
       {profileLinks}
-    </Wrapper>
+    </>
   );
 };
